The TestService is now a standalone function called equals.
The parameters in the test service were called got and want.
This is unnecessary information as this compare function can compare any list of documents.
So I renamed them to a and b.
This is ok since the functions are so short.

I renamed "compare" to "equals" because equals is a known function name in Java.
Maybe the existing equals also does this job.
THen we would not need the custom one.